refactor: consolidate permissionscope components in own pkg
