+++ pkg/azure/client/application/approles_test.go
@@ -1,4 +1,4 @@
-package approle_test
+package application_test
 
 import (
 	"testing"
