+++ pkg/azure/client/application/permissionscope/const.go
