+++ pkg/azure/client/application/permissionscope/interface.go
