+++ pkg/azure/client/application/permissionscope/permissionscope.go
