+++ pkg/azure/client/application/permissionscope/permissionscopemap.go
