+++ pkg/azure/client/application/permissionscope/result.go
