+++ pkg/azure/util/permissions/permissions_test.go
@@ -10,8 +10,8 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/nais/azureator/pkg/azure/client/application/approle"
+	"github.com/nais/azureator/pkg/azure/client/application/permissionscope"
 	"github.com/nais/azureator/pkg/azure/util/permissions"
-	"github.com/nais/azureator/pkg/azure/util/permissionscope"
 )
 
 func TestGenerateDesiredPermissionSet(t *testing.T) {
